requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
django-tinymce==3.5.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
